Pull thank you modal logic up one level
