+++ src/components/pages/educators/sign-up-modal.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { size, lineHeight } from '../../dev-hub/theme';
 import { H5, P } from '../../dev-hub/text';
 import styled from '@emotion/styled';
@@ -6,6 +6,7 @@ import { css } from '@emotion/core';
 import Button from '../../dev-hub/button';
 import AcademiaSignUpForm from '../../dev-hub/academia-sign-up-form';
 import Modal from '../../dev-hub/modal';
+import SuccessState from '~components/dev-hub/success-state';
 
 const MAX_SIGN_UP_WIDTH = '600px';
 
@@ -22,24 +23,38 @@ const StyledAcademiaSignUpForm = styled(AcademiaSignUpForm)`
     margin-top: ${size.large};
 `;
 
-const SignUpModal = () => (
-    <Modal
-        triggerComponent={
-            <ButtonWithAdditionalTopMargin primary hasArrow={false}>
-                Join MongoDB for Academia
-            </ButtonWithAdditionalTopMargin>
-        }
-        dialogContainerStyle={{ maxWidth: MAX_SIGN_UP_WIDTH }}
-    >
-        <H5>Join MongoDB for Academia</H5>
-        <P css={defaultLineHeight}>
-            If you’re interested in receiving MongoDB course materials or if you
-            like us to review your current content, please let us know by
-            submitting the form and we’ll get back to you within five business
-            days.
-        </P>
-        <StyledAcademiaSignUpForm />
-    </Modal>
-);
+const StyledSuccessState = styled(SuccessState)`
+    margin-bottom: ${size.xlarge};
+    margin-top: ${size.xlarge};
+`;
+
+const SignUpModal = () => {
+    const [success, setSuccess] = useState(null);
+    if (success) {
+        return <StyledSuccessState>Thank you for joining!</StyledSuccessState>;
+    }
+    return (
+        <Modal
+            triggerComponent={
+                <ButtonWithAdditionalTopMargin primary hasArrow={false}>
+                    Join MongoDB for Academia
+                </ButtonWithAdditionalTopMargin>
+            }
+            dialogContainerStyle={{ maxWidth: MAX_SIGN_UP_WIDTH }}
+        >
+            <H5>Join MongoDB for Academia</H5>
+            <P css={defaultLineHeight}>
+                If you’re interested in receiving MongoDB course materials or if
+                you like us to review your current content, please let us know
+                by submitting the form and we’ll get back to you within five
+                business days.
+            </P>
+            <StyledAcademiaSignUpForm
+                setSuccess={setSuccess}
+                success={success}
+            />
+        </Modal>
+    );
+};
 
 export default SignUpModal;
